leam-base.commons-lang.mid-411.idx-40259.22.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableFloat.java
@@ -266,11 +266,10 @@ public class MutableFloat extends Number implements Comparable<MutableFloat>, Mu
      * @return the value associated with this instance immediately before the operand was added
      * @since 3.5
      */
-    public float getAndAdd(final float operand) {
-        final float last = value;
-        this.value += operand;
-        return last;
-    }
+        public float getAndAdd(final float operand) {
+            final float last = value;
+            this.value += operand;
+    return (float)last;    }
 
     /**
      * Increments this instance's value by {@code operand}; this method returns the value associated with the instance
